Add checks for the content of the created files.
The check for the leap-seconds in /etc/ntp.leapseconds is not strictly necessary but serves as a cross-check for create_file_with_content.

Check for .*Chef.* in /etc/ntp.conf to see if chef created the file from template.
Check for servers in /etc/ntp.conf to see if the servers end up where they should (currently fails with some template error).
Check for special servers and peers (currently fails with the same template error).